[RF] Avoid including `RooFactoryWSTool` in public interfaces
